Use a ViewModel for the Dynamic Type example, move classes and extensions out of the ViewController.
